feature: add not enough memory warning screen (#96)
* Feature: Warning for high or low max allocated RAM

Added a method which checks the ram allocation and then gives a warning if it is offset by 1100 MB from the recommended max ram allocation. The offset for 1100 is set because AA-like categories use about 1000 MB more than other categories. MultiMC's Ram Allocation amounts may pass the 1000 MB barrier, to avoid it we add 100 more to the 1000 MB. There's also a system property to disable it.

* feature: warning screen for allocated memory check

Replaced the log message with a screen, removed the upper memory limit.

* feat: check memory when starting resetting

- Change the low memory warning to trigger when starting resetting,
not on launch, as it depends on the max queued seeds config.

- Mention max queued seeds in the warning, and link to the wiki.

* feat: memory warning wiki link description

Link: https://github.com/contariaa/seedqueue/pull/96

---------

Co-authored-by: Crystal15118 <138210704+Crystal15118@users.noreply.github.com>